Struct nvapi_sys::gpu::display::NV_GPU_DISPLAYIDS

#[repr(C)]

pub struct NV_GPU_DISPLAYIDS {
    pub version: u32,
    pub connectorType: NV_MONITOR_CONN_TYPE,
    pub displayId: u32,
    pub flags: NV_GPU_DISPLAYIDS_FLAGS,
}

Fields
version: u32 connectorType: NV_MONITOR_CONN_TYPE

out: vga, tv, dvi, hdmi and dp. This is reserved for future use and clients should not rely on this information. Instead get the GPU connector type from NvAPI_GPU_GetConnectorInfo/NvAPI_GPU_GetConnectorInfoEx

displayId: u32

this is a unique identifier for each device

flags: NV_GPU_DISPLAYIDS_FLAGS

if bit is set then this display is part of MST topology and it's a dynamic

Methods

impl NV_GPU_DISPLAYIDS

pub fn zeroed() -> Self