Struct nvapi_sys::gpu::power::private::PerfFlags

pub struct PerfFlags { /* fields omitted */ }

Methods

impl PerfFlags

const POWER_LIMIT: PerfFlags

POWER_LIMIT: PerfFlags = PerfFlags{bits: 1,}

const TEMPERATURE_LIMIT: PerfFlags

TEMPERATURE_LIMIT: PerfFlags = PerfFlags{bits: 2,}

const VOLTAGE_LIMIT: PerfFlags

VOLTAGE_LIMIT: PerfFlags = PerfFlags{bits: 4,}

const UNKNOWN_8: PerfFlags

UNKNOWN_8: PerfFlags = PerfFlags{bits: 8,}

Seen once with driver crash where flags = 0x0f

const NO_LOAD_LIMIT: PerfFlags

NO_LOAD_LIMIT: PerfFlags = PerfFlags{bits: 16,}

const UNKNOWN_32: PerfFlags

UNKNOWN_32: PerfFlags = PerfFlags{bits: 32,}

Never seen this

fn empty() -> PerfFlags

Returns an empty set of flags.

fn all() -> PerfFlags

Returns the set containing all flags.

fn bits(&self) -> NV_GPU_PERF_FLAGS

Returns the raw value of the flags currently stored.

fn from_bits(bits: NV_GPU_PERF_FLAGS) -> Option<PerfFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: NV_GPU_PERF_FLAGS) -> PerfFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: PerfFlags) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: PerfFlags) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: PerfFlags)

Inserts the specified flags in-place.

fn remove(&mut self, other: PerfFlags)

Removes the specified flags in-place.

fn toggle(&mut self, other: PerfFlags)

Toggles the specified flags in-place.

fn set(&mut self, other: PerfFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.