Module nvapi_sys::gpu::thermal

The GPU thermal control APIs are used to get temperature levels from the various thermal sensors associated with the GPU.

Modules

private
Undocumented API

Structs

NV_GPU_THERMAL_SETTINGS_SENSOR
Anonymous struct in NV_GPU_THERMAL_SETTINGS

NV_GPU_THERMAL_SETTINGS_V1
Used in NvAPI_GPU_GetThermalSettings()

Enums

ThermalController
NV_GPU_THERMAL_SETTINGS

ThermalTarget
Used in NV_GPU_THERMAL_SETTINGS

Constants

NVAPI_MAX_THERMAL_SENSORS_PER_GPU

NVAPI_THERMAL_CONTROLLER_ADM1032

NVAPI_THERMAL_CONTROLLER_ADT7473

NVAPI_THERMAL_CONTROLLER_GPU_INTERNAL

NVAPI_THERMAL_CONTROLLER_LM64

NVAPI_THERMAL_CONTROLLER_LM89

NVAPI_THERMAL_CONTROLLER_LM99

NVAPI_THERMAL_CONTROLLER_MAX1617

NVAPI_THERMAL_CONTROLLER_MAX6649

NVAPI_THERMAL_CONTROLLER_NONE

NVAPI_THERMAL_CONTROLLER_OS

NVAPI_THERMAL_CONTROLLER_SBMAX6649

NVAPI_THERMAL_CONTROLLER_UNKNOWN

NVAPI_THERMAL_CONTROLLER_VBIOSEVT

NVAPI_THERMAL_TARGET_ALL

NVAPI_THERMAL_TARGET_BOARD
GPU board ambient temperature requires NvPhysicalGpuHandle

NVAPI_THERMAL_TARGET_GPU
GPU core temperature requires NvPhysicalGpuHandle

NVAPI_THERMAL_TARGET_MEMORY
GPU memory temperature requires NvPhysicalGpuHandle

NVAPI_THERMAL_TARGET_NONE

NVAPI_THERMAL_TARGET_POWER_SUPPLY
GPU power supply temperature requires NvPhysicalGpuHandle

NVAPI_THERMAL_TARGET_UNKNOWN

NVAPI_THERMAL_TARGET_VCD_BOARD
Visual Computing Device Board temperature requires NvVisualComputingDeviceHandle

NVAPI_THERMAL_TARGET_VCD_INLET
Visual Computing Device Inlet temperature requires NvVisualComputingDeviceHandle

NVAPI_THERMAL_TARGET_VCD_OUTLET
Visual Computing Device Outlet temperature requires NvVisualComputingDeviceHandle

NV_GPU_THERMAL_SETTINGS_VER

NV_GPU_THERMAL_SETTINGS_VER_1

NV_GPU_THERMAL_SETTINGS_VER_2

Functions

NvAPI_GPU_GetThermalSettings⚠
This function retrieves the thermal information of all thermal sensors or specific thermal sensor associated with the selected GPU.

Type Definitions

GPU_GetThermalSettingsFn

NV_GPU_THERMAL_SETTINGS

NV_GPU_THERMAL_SETTINGS_V2

NV_THERMAL_CONTROLLER
NV_GPU_THERMAL_SETTINGS

NV_THERMAL_TARGET
Used in NV_GPU_THERMAL_SETTINGS