Module nvapi_sys::handles

NVAPI Handles - These handles are retrieved from various calls and passed in to others in NvAPI These are meant to be opaque types. Do not assume they correspond to indices, HDCs, display indexes or anything else.

Most handles remain valid until a display re-configuration (display mode set) or GPU reconfiguration (going into or out of SLI modes) occurs. If NVAPI_HANDLE_INVALIDATED is received by an app, it should discard all handles, and re-enumerate them.

Structs

NVDX_SwapChainHandle
DirectX SwapChain objects

Nv3DVPContextHandle
A handle for a 3D Vision Pro (3DVP) context

Nv3DVPGlassesHandle
A handle for a pair of 3DVP RF shutter glasses

Nv3DVPTransceiverHandle
A handle for a 3DVP RF transceiver

NvAudioHandle
NVIDIA HD Audio Device

NvDisplayHandle
Display Device driven by NVIDIA GPU(s) (an attached display)

NvEventHandle
A handle to an event registration instance

NvGSyncDeviceHandle
A handle to a Sync device

NvHICHandle
A handle to a Host Interface Card

NvLogicalGpuHandle
One or more physical GPUs acting in concert (SLI)

NvMonitorHandle
Monitor handle

NvPhysicalGpuHandle
A single physical GPU

NvSourceHandle
Unique source handle on the system

NvTargetHandle
Unique target handle on the system

NvTransitionHandle
A handle to address a single transition request

NvUnAttachedDisplayHandle
Unattached Display Device driven by NVIDIA GPU(s)

NvVioHandle
A handle to an SDI device

NvVisualComputingDeviceHandle
A handle to a Visual Computing Device

Constants

NVAPI_DEFAULT_HANDLE

NVDX_SWAPCHAIN_NONE

Type Definitions

StereoHandle
A stereo handle, that corresponds to the device interface